+++ src/Atoms/Icon/IconBase.tsx
@@ -5,22 +5,23 @@ import { InternalProps, StyledIconBaseProps } from './IconBase.types';
 const StyledIconBase = styled.svg<StyledIconBaseProps>`
   ${p => {
     const size = p.size || 5;
+    const fill = p.colorFn ? p.colorFn(p.theme) : p.theme.color.svgFill;
 
-    return ` user-select: none;
-    width: ${p.theme.spacing.unit(size)}px;
-    height: ${p.theme.spacing.unit(size)}px;
-    fill: currentColor;
-    flex-shrink: 0;
-    color: ${p.colorFn ? p.colorFn(p.theme) : p.theme.color.text};
+    return `
+      user-select: none;
+      width: ${p.theme.spacing.unit(size)}px;
+      height: ${p.theme.spacing.unit(size)}px;
+      fill: ${fill};
+      flex-shrink: 0;
 
-    display: ${p.inline ? 'inline-block' : 'block'};
-    ${p.inline ? 'vertical-align: middle;' : ''}
+      display: ${p.inline ? 'inline-block' : 'block'};
+      ${p.inline ? 'vertical-align: middle;' : ''}
     `;
   }}
 `;
 
 export const IconBase: React.FC<InternalProps> = (props: InternalProps) => {
-  const { className, children, title, size = 5, color, inline } = props;
+  const { className, children, title, size = 5, color, fill, inline } = props;
 
   return (
     <StyledIconBase
@@ -30,7 +31,7 @@ export const IconBase: React.FC<InternalProps> = (props: InternalProps) => {
       viewBox="0 0 24 24"
       aria-hidden={title ? 'false' : 'true'}
       role={title ? 'img' : 'presentation'}
-      colorFn={color}
+      colorFn={fill || color}
       inline={inline}
     >
       {children}