+++ src/Atoms/Icon/IconBase.types.ts
@@ -10,14 +10,24 @@ export type StyledIconBaseProps = {
   inline?: boolean;
 };
 
+export type StyledChildProps = {
+  strokeColorFn?: ColorFn;
+};
+
 export type BaseProps = {
   className?: string;
   title?: string;
+  /** @deprecated use fill instead */
   color?: ColorFn;
+  fill?: ColorFn;
   /** unit-based */
   size?: number;
   inline?: boolean;
 };
+
+export type ChildProps = {
+  stroke?: ColorFn;
+} & BaseProps;
 
 export type InternalProps = BaseProps & {
   children: React.ReactNode;