+++ src/index.ts
@@ -1,5 +1,6 @@
 /** Atoms */
 import { Table, Tbody, Td, Tfoot, Th, Thead, Tr } from './Atoms/Table';
+import Box from './Atoms/Box';
 import Card from './Atoms/Card';
 import CssGrid from './Atoms/CssGrid';
 import Flag from './Atoms/Flag';
@@ -13,7 +14,6 @@ import Spinner from './Atoms/Spinner';
 import TabTitle from './Atoms/TabTitle';
 import Time from './Atoms/Time';
 import Typography from './Atoms/Typography';
-import Box from './Atoms/Box';
 import VisuallyHidden from './Atoms/VisuallyHidden';
 
 /** Molecules */
@@ -30,6 +30,7 @@ import Number from './Molecules/Number';
 import PageHeaderCard from './Molecules/PageHeaderCard';
 import PageWrapper from './Molecules/PageWrapper';
 import Select from './Molecules/Select';
+import Switch from './Molecules/Switch';
 import Tabs from './Molecules/Tabs';
 import TabsNav from './Molecules/TabsNav';
 
@@ -40,13 +41,14 @@ import Development from './Organisms/Development';
 import theme, { createTheme } from './theme';
 /** Runtime code */
 export {
-  createTheme,
   Avatar,
+  Box,
   Button,
   Card,
   CardWithTabs,
   CardWithTitle,
   CollapsibleCard,
+  createTheme,
   CssGrid,
   Development,
   Flag,
@@ -67,6 +69,7 @@ export {
   Select,
   Separator,
   Spinner,
+  Switch,
   Table,
   Tabs,
   TabsNav,
@@ -80,6 +83,5 @@ export {
   Time,
   Tr,
   Typography,
-  Box,
   VisuallyHidden,
 };